refactor: lesson Dto 수정
