+++ server/src/main/java/com/codueon/boostUp/domain/lesson/controller/LessonController.java
@@ -10,7 +10,7 @@ import com.codueon.boostUp.domain.lesson.dto.patch.PostLessonDetailEdit;
 import com.codueon.boostUp.domain.lesson.dto.patch.PostLessonInfoEdit;
 import com.codueon.boostUp.domain.lesson.dto.post.PostLesson;
 import com.codueon.boostUp.domain.lesson.dto.post.PostSearchLesson;
-import com.codueon.boostUp.domain.lesson.dto.utils.WrapLessonId;
+import com.codueon.boostUp.domain.lesson.dto.get.GetLessonId;
 import com.codueon.boostUp.domain.lesson.service.LessonService;
 import com.codueon.boostUp.global.security.token.JwtAuthenticationToken;
 import lombok.RequiredArgsConstructor;
@@ -124,11 +124,11 @@ public class LessonController {
      * @author Quartz614
      */
     @GetMapping(value = "/tutor")
-    public ResponseEntity<WrapLessonId> getLessonMypage(Authentication authentication) {
+    public ResponseEntity<GetLessonId> getLessonMypage(Authentication authentication) {
         JwtAuthenticationToken token = (JwtAuthenticationToken) authentication;
         Long memberId = getMemberIdIfExistToken(token);
 
-        return ResponseEntity.ok().body(new WrapLessonId(lessonService.getLessonMypage(memberId)));
+        return ResponseEntity.ok().body(new GetLessonId(lessonService.getLessonMypage(memberId)));
     }
 
     /**
